requirements.txt
streamlit
requests
Pillow
numpy
opencv-python
python-dotenv